07 1-column-chart; code style fixes, helpers have been moved to a common project folder
