+++ 05-dom-document-loading/2-sortable-table-v1/templates/body-custom-cell-template.js
@@ -1,4 +1,4 @@
-import { generateElement } from "../helpers/element-helper.js";
+import { generateElement } from "../../../common/helpers/element-helper.js";
 
 export class BodyCustomCellTemplate {
   constructor(data = [], templateRender = () => { }) {
