+++ 05-dom-document-loading/2-sortable-table-v1/templates/header-template.js
@@ -1,5 +1,5 @@
 import { HeaderCellTemplate } from "./header-cell-template.js";
-import { generateElement } from "../helpers/element-helper.js";
+import { generateElement } from "../../../common/helpers/element-helper.js";
 
 export class HeaderTemplate {
   constructor(headerConfig = [], sorted = {}) {
