+++ common/helpers/comparer-helper.js
