+++ common/helpers/element-helper.js
